StipplePlotly

Missing docstring.

Missing docstring for Font. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ColorBar. Check Documenter's build log for details.

Base.Dict — Type

Dict([itr])

Dict{K,V}() constructs a hash table with keys of type K and values of type V. Keys are compared with isequal and hashed with hash.

Given a single iterable argument, constructs a Dict whose key-value pairs are taken from 2-tuples (key,value) generated by the argument.

Examples

julia> Dict([("A", 1), ("B", 2)])
Dict{String, Int64} with 2 entries:
  "B" => 2
  "A" => 1

Alternatively, a sequence of pair arguments may be passed.

julia> Dict("A"=>1, "B"=>2)
Dict{String, Int64} with 2 entries:
  "B" => 2
  "A" => 1

Missing docstring.

Missing docstring for optionals!. Check Documenter's build log for details.

Stipple.render — Function

`function render`

Abstract function. Needs to be specialized by plugins. It is automatically invoked by Stipple to serialize a Julia data type (corresponding to the fields in the ReactiveModel instance) to JavaScript/JSON. In general the specialized methods should return a Julia Dict which are automatically JSON encoded by Stipple. If custom JSON serialization is required for certain types in the resulting Dict, specialize JSON.lower for that specific type.

Example

function Stipple.render(ps::PlotSeries, fieldname::Union{Symbol,Nothing} = nothing)
  Dict(:name => ps.name, ps.plotdata.key => ps.plotdata.data)
end

Specialized JSON rendering for Undefined

JSON.lower(x::Undefined) = "__undefined__"

Missing docstring.

Missing docstring for ErrorBar. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlotAnnotation. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlotLayoutGrid. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlotLayoutAxis. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlotLayoutTitle. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlotLayoutLegend. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlotLayout. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlotlyLine. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlotDataMarker. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlotData. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlotConfig. Check Documenter's build log for details.

Missing docstring.

Missing docstring for plot. Check Documenter's build log for details.